fix(review): address final PR 1986 cleanup
