+++ shared/src/index.ts
@@ -93,7 +93,6 @@ export * from "./utils/calorieCalculations.ts";
 export * from "./utils/servingSizeConversions.ts";
 export * from "./utils/nutrientMatching.ts";
 export * from "./utils/search.ts";
-export * from "./utils/workoutSource.ts";
 export * from "./ai/unitConversion.ts";
 export * from "./ai/confidenceLabels.ts";
 export * from "./medications/contracts.ts";
